Database restructure to include Courses and field refractoring The courses model was added to the database and the fields where changed to align with the Canvas API.

+++ grader/views.py
@@ -11,20 +11,17 @@ from rest_framework.exceptions import AuthenticationFailed
 from rest_framework.authentication import SessionAuthentication, BasicAuthentication
 from rest_framework.permissions import IsAuthenticated
 import jwt, datetime
-from .models import File
-from .autograde import autograde
+from .models import Course, File, Assignment
 from django.utils.decorators import method_decorator
 from django.core import serializers
 from django.views.decorators.csrf import csrf_exempt
-from .serializers import QuestionSerializer, FileSerializer, TestCaseSerializer
+from .serializers import AssignmentSerializer, CourseSerializer, FileSerializer, TestCaseSerializer
 from home.models import User
-from grader.models import TestCase, Question, File
+from grader.models import TestCase, Assignment, File
 import random
 import string
 import os
 
-
-# Create your views here.
 
 def slug_generator(size=6, chars=string.ascii_uppercase + string.digits):
     return ''.join(random.choice(chars) for _ in range(size))
@@ -48,8 +45,9 @@ def validate_user(request):
 
     return True, user
 
+
 @method_decorator(csrf_exempt, name='dispatch')
-class QuestionView(APIView):
+class CourseView(APIView):
 
     def get(self, request):
         stat, user = validate_user(request)
@@ -59,31 +57,16 @@ class QuestionView(APIView):
                 "message": "no user found"
             }, status=status.HTTP_401_UNAUTHORIZED)
 
-        if not user.is_lecturer:
-            return Response({
-                "status": "error",
-                "message": "only lecturers can upload a question"
-            }, status=status.HTTP_401_UNAUTHORIZED)
-
-        quest = Question.objects.filter(lecturer=user)
-        question = QuestionSerializer(quest, many=True)
+        cours = Course.objects.all()
+        course = CourseSerializer(cours, many=True)
         return Response({
             "status": "success",
-            "data": question.data,
-            "message": "Question added"
+            "data": course.data,
+            "message": "Assignment query succesful"
         }, status=status.HTTP_200_OK)
 
 
     def post(self, request):
-
-        # if not request.user.is_authenticated:
-        #     return Response({
-        #         "status": "error",
-        #         "message": "no user found"
-        #     }, status=status.HTTP_401_UNAUTHORIZED)
-
-        # user = request.user
-
         stat, user = validate_user(request)
         if not stat:
             return Response({
@@ -94,36 +77,16 @@ class QuestionView(APIView):
         if not user.is_lecturer:
             return Response({
                 "status": "error",
-                "message": "only lecturers can upload a question"
+                "message": "only lecturers can create a course"
             }, status=status.HTTP_401_UNAUTHORIZED)
 
-        question = QuestionSerializer(data=request.data)
-        question.is_valid(raise_exception=True)
-
-        slug = slug_generator()
-        question.validated_data['lecturer'] = user
-        question.validated_data['slug'] = slug
-        question.save()
-        quest = Question.objects.get(slug=slug)
-        # upload = TestCaseSerializer(data=request.FILES)
-        # if upload.is_valid():
-        for f in request.FILES.getlist('file'):
-            TestCase.objects.create(file=f, question=quest)
-        # else:
-        #     print("not valid")
-        #     print(request.FILES)
-
-        from autograderstable.autograder.autograder import AutograderPaths, Grader
-        from autograderstable.autograder import guide
-
-        current_dir = f'media/upload/{quest.slug}'
-        os.mkdir(f"{current_dir}/")
-        os.mkdir(f"{current_dir}/results")
-        guide.main(AutograderPaths(current_dir))
+        course = CourseSerializer(data=request.data)
+        course.is_valid(raise_exception=True)
+        course.save()
 
         return Response({
             "status": "success",
-            "data": question.data,
+            "data": course.data,
             "message": "Question added"
         }, status=status.HTTP_200_OK)
 
@@ -141,7 +104,93 @@ class QuestionView(APIView):
                 "message": "You are not a lecturer"
             }, status=status.HTTP_401_UNAUTHORIZED)
 
-        question = get_object_or_404(Question, id=id)
+        course = get_object_or_404(Course, id=id)
+        course.delete()
+
+        return Response({
+            "status": "success",
+            "message": "Course deleted"
+        }, status=status.HTTP_200_OK)
+
+
+
+@method_decorator(csrf_exempt, name='dispatch')
+class AssignmentView(APIView):
+
+    def get(self, request):
+        stat, user = validate_user(request)
+        course = request.data['course_id']
+        if not stat:
+            return Response({
+                "status": "error",
+                "message": "no user found"
+            }, status=status.HTTP_401_UNAUTHORIZED)
+
+        quest = Assignment.objects.filter(course_id=course)
+        question = AssignmentSerializer(quest, many=True)
+        return Response({
+            "status": "success",
+            "data": question.data,
+            "message": "Assignment query succesful"
+        }, status=status.HTTP_200_OK)
+
+
+    def post(self, request):
+        stat, user = validate_user(request)
+        course = request.data['course_id']
+        if not stat:
+            return Response({
+                    "status": "error",
+                    "message": "no user found"
+            }, status=status.HTTP_401_UNAUTHORIZED)
+
+        if not user.is_lecturer:
+            return Response({
+                "status": "error",
+                "message": "only lecturers can create an assignment"
+            }, status=status.HTTP_401_UNAUTHORIZED)
+
+        question = AssignmentSerializer(data=request.data)
+        question.is_valid(raise_exception=True)
+
+        slug = slug_generator()
+        question.validated_data['course_id'] = course
+        question.validated_data['slug'] = slug
+        question.save()
+        quest = Assignment.objects.get(slug=slug)
+
+        for f in request.FILES.getlist('file'):
+            TestCase.objects.create(file=f, question=quest)
+
+        from autograderstable.autograder.autograder import AutograderPaths, Grader
+        from autograderstable.autograder import guide
+
+        current_dir = f'media/upload/{quest.slug}'
+        os.mkdir(f"{current_dir}/")
+        os.mkdir(f"{current_dir}/results")
+        guide.main(AutograderPaths(current_dir))
+
+        return Response({
+            "status": "success",
+            "data": question.data,
+            "message": "Assignment added"
+        }, status=status.HTTP_200_OK)
+
+    def delete(self, request, id):
+        stat, user = validate_user(request)
+        if not stat:
+            return Response({
+                "status": "error",
+                "message": "no user found"
+            }, status=status.HTTP_401_UNAUTHORIZED)
+
+        if not user.is_lecturer:
+            return Response({
+                "status": "error",
+                "message": "You are not a lecturer"
+            }, status=status.HTTP_401_UNAUTHORIZED)
+
+        question = get_object_or_404(Assignment, id=id)
 
         if os.path.exists(f"media/upload/{question.slug}"):
             shutil.rmtree(f"media/upload/{question.slug}")
@@ -149,25 +198,24 @@ class QuestionView(APIView):
 
         return Response({
             "status": "success",
-            "message": "Question deleted"
+            "message": "Assignment deleted"
         }, status=status.HTTP_200_OK)
 
 class TestCaseView(APIView):
 
     def post(self, request):
-        id = request.data['question']
-        quest = Question.objects.get(id=id)
+        assignment_id = request.data['assignment_id']
         testcases = {}
         num = 1
         for f in request.FILES.getlist('file'):
-            test = TestCase.objects.create(file=f, question=quest)
+            test = TestCase.objects.create(file=f, assignment_id=assignment_id)
             testcases[num] = TestCaseSerializer(test).data
             num +=1
 
         return Response({
             "status": "success",
             "data": testcases,
-            "message": "Question added"
+            "message": "Testcase added"
         }, status=status.HTTP_200_OK)
 
     def get(self, request, id):
@@ -184,14 +232,14 @@ class TestCaseView(APIView):
                 "message": "You are not a lecturer"
             }, status=status.HTTP_401_UNAUTHORIZED)
 
-        question = Question.objects.get(id=id, lecturer=user)
+        question = Assignment.objects.get(id=id, lecturer=user)
         if not question:
             return Response({
                 "status": "success",
                 "message": "No question with id"
             }, status=status.HTTP_404_NOT_FOUND)
 
-        testcase = TestCase.objects.filter(question=question)
+        testcase = TestCase.objects.filter(assignment_id=question)
         testcaseserializer = TestCaseSerializer(testcase, many=True)
 
         return Response({
@@ -227,27 +275,29 @@ class TestCaseView(APIView):
             "message": "Test case deleted"
         }, status=status.HTTP_200_OK)
 
-class FileView(APIView):
+class SubmissionView(APIView):
     parser_classes = (MultiPartParser, FormParser)
 
     def post(self, request, *args, **kwargs):
+        stat, user = validate_user(request)
+        if not stat:
+            return Response({
+                "status": "error",
+                "message": "no user found"
+            }, status=status.HTTP_401_UNAUTHORIZED)
 
-        file_serializer = FileSerializer(data=request.data)
-
-        if not file_serializer.is_valid():
-            return Response(file_serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-
-        file_serializer.save()
-        id = file_serializer.data['id']
-
+        f = request.data['file']
+        assignment_id = request.data['assignment_id']
+        sub = File.objects.create(file=f, user_id=user.id, assignment_id=assignment_id)
+        file_sub = FileSerializer(sub).data
+        id = file_sub['id']
+       
         file = File.objects.get(id=id)
-
         # run the autograder
-        os.system(f"python ./tester.py 'media/upload/{file.question.slug}' '{file.id}'")
+        os.system(f"python ./tester.py 'media/upload/{file.assignment.slug}' '{file.id}'")
 
         file = File.objects.get(id=id)
         file_serializer = FileSerializer(file)
-        # obj = serializers.serialize('json', [file,])
 
         if os.path.exists(f"media/{file.file}"):
             os.remove(f"media/{file.file}")
@@ -256,6 +306,32 @@ class FileView(APIView):
             "status": "success",
             "data": file_serializer.data,
         }, status=status.HTTP_201_CREATED)
+
+    def get(self, request):
+
+        stat, user = validate_user(request)
+        if not stat:
+            return Response({
+                "status": "error",
+                "message": "no user found"
+            }, status=status.HTTP_401_UNAUTHORIZED)
+
+        question = File.objects.get(user_id=user)
+        if not question:
+            return Response({
+                "status": "success",
+                "message": "No question with id"
+            }, status=status.HTTP_404_NOT_FOUND)
+
+
+        files = File.objects.filter(assignment_id=question)
+        file_serializer = FileSerializer(files, many=True)
+
+        return Response({
+            "status": "success",
+            "data": file_serializer.data,
+            "message": "Question added"
+        }, status=status.HTTP_200_OK)
 
     def get(self, request, id):
 
@@ -272,21 +348,20 @@ class FileView(APIView):
                 "message": "Unauthorized"
             }, status=status.HTTP_401_UNAUTHORIZED)
 
-        question = Question.objects.get(id=id, lecturer=user)
+        question = File.objects.get(id=id)
         if not question:
             return Response({
                 "status": "success",
-                "message": "No question with id"
+                "message": "No submission with id"
             }, status=status.HTTP_404_NOT_FOUND)
 
 
-        files = File.objects.filter(question=question)
+        files = File.objects.filter(assignment_id=question)
         file_serializer = FileSerializer(files, many=True)
 
         return Response({
             "status": "success",
             "data": file_serializer.data,
-            "message": "Question added"
+            "message": "Submission query completed"
         }, status=status.HTTP_200_OK)
-
